compare latest blessed version for bytewise equality (#36)
The Dropshot API manager detects semantic equality using
[drift](https://docs.rs/drift). This was primarily intended so that
wire-compatible changes could be made to endpoint handlers for old
blessed versions. But, as a side-effect, this property also applies to
the latest blessed version. This means that trivial changes to blessed
versions can accumulate, up to the point at which a non-trivial change
happens, and then the diff becomes quite confusing.

Address this by special-casing the latest blessed version, and requiring
both bytewise and semantic equality for it. Also allow
client-side-versioned APIs (which are currently frozen; see
https://github.com/oxidecomputer/omicron/issues/9290) to opt out of this
scheme.

Closes #35.